refactor(CustomComboBox): rename `types.ts`
